pyproject.toml
[project]
name = "pyview-map"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "pyview-web>=v0.9.0a4",
    "uvicorn>=0.41.0",
    "fastapi>=0.115",
    "httpx>=0.28",
    "http-stream-transport",
    "http-stream-client",
    "jrpc-common",
]

[tool.uv.sources]
http-stream-transport = { git = "ssh://git@github.com/avilior/http_stream_prj", subdirectory = "packages/server_pkg" }
http-stream-client = { git = "ssh://git@github.com/avilior/http_stream_prj", subdirectory = "packages/client" }
jrpc-common = { git = "ssh://git@github.com/avilior/http_stream_prj", subdirectory = "packages/jrpc_common" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pyview_map"]

[tool.ruff]
line-length = 120
# Ruff doesn't support t-strings (PEP 750, Python 3.14) yet — exclude files that use them
exclude = [
    "src/pyview_map/components/dynamic_map/dynamic_map_component.py",
    "src/pyview_map/components/dynamic_map/map_driver.py",
    "src/pyview_map/components/dynamic_list/dynamic_list.py",
    "src/pyview_map/components/dynamic_list/list_driver.py",
    "src/pyview_map/applications/flights_demo/flights_demo.py",
    "src/pyview_map/applications/places_demo/places_demo.py",
]

[tool.ruff.format]
skip-magic-trailing-comma = true

[project.scripts]
pyview-map = "pyview_map.__main__:main"
